engine:: MGStreamLevel: pass array of additional imports using MIMMap
