+++ engine/src/main/java/good/damn/engine2/level/MGIProviderMapImport.kt
@@ -1,0 +1,10 @@
+package good.damn.engine2.level
+
+import good.damn.mapimporter.models.MIMMap
+
+interface MGIProviderMapImport {
+    fun import(
+        map: MIMMap,
+        localPathDir: String
+    )
+}
